[refactor][frontend] Refatorando e organizando melhor os arquivos do frontend
