+++ frontend/src/app/app.component.html
@@ -1,10 +1,4 @@
-<main class="main">
-  <header class="brand-name">
-    <button class="logout" (click)="logout()" *ngIf="isAuthenticated()">
-      Logout
-    </button>
-  </header>
-  <section class="content">
-    <router-outlet></router-outlet>
-  </section>
-</main>
+<section class="content">
+  <router-outlet></router-outlet>
+</section>
+
