+++ frontend/src/app/home/home.component.html
@@ -1,0 +1,10 @@
+<main class="main">
+  <header class="brand-name">
+    <button class="logout" (click)="logout()" *ngIf="isAuthenticated()">
+      Logout
+    </button>
+  </header>
+  <section class="content">
+    <p class="welcome-message">Bem vindo!</p>
+  </section>
+</main>
