--- frontend/src/app/home/home.html
@@ -1,1 +1,0 @@
-<p>home works!</p>
